backend/requirements.txt
livekit-agents[anthropic,deepgram,cartesia,silero]>=0.12.0
livekit>=0.17.0
livekit-api>=0.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
twilio>=9.4.0
httpx>=0.27.0
pydantic>=2.7.0
websockets>=12.0
aiofiles>=23.2.1
